// apps/web/src/lib/api/models/TaskCommentResponseDto.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
import type { UserInfoDto } from './UserInfoDto';
export type TaskCommentResponseDto = {
    id: string;
    taskId: string;
    author: UserInfoDto;
    content: string;
    createdAt: string;
    updatedAt: string;
};